requirements.txt
click==7.1.2 
numpy==1.21.4 
pandas==1.1.4 
pillow==8.0.0 
scikit-learn==0.23.2 
#torch==1.7.0 
torch==1.12.1
#torchvision==0.8.1
torchvision==0.13.1
torchaudio==0.12.1
